+++ TestTool/TransMsg.cpp
@@ -1,6 +1,6 @@
 #include "TransMsg.h"
 
-void showMfcMsg(const char *pszMsg, const size_t iLens)
+void showMfcMsg(const char *pszMsg, const unsigned int iLens)
 {
     char *pBuff = new char[iLens + 1];
     memset(pBuff, 0, iLens + 1);
@@ -9,7 +9,7 @@ void showMfcMsg(const char *pszMsg, const size_t iLens)
     SendMessage(g_hWnd, WM_MESSAGE_SHOWMFCMSG, (WPARAM)pBuff, iLens);
 }
 
-void showLuaMemory(const char *pszMsg, const size_t iLens)
+void showLuaMemory(const char *pszMsg, const unsigned int iLens)
 {
     char *pBuff = new char[iLens + 1];
     memset(pBuff, 0, iLens + 1);
