+++ TestTool/TransMsg.h
@@ -7,8 +7,8 @@
 #define WM_MESSAGE_ENBLELINKBUTT WM_MESSAGE_SHOWMFCMSG + 1
 #define WM_MESSAGE_SHOWLUAMEMORY WM_MESSAGE_ENBLELINKBUTT + 1
 
-void showMfcMsg(const char *pszMsg, const size_t iLens);
-void showLuaMemory(const char *pszMsg, const size_t iLens);
+void showMfcMsg(const char *pszMsg, const unsigned int iLens);
+void showLuaMemory(const char *pszMsg, const unsigned int iLens);
 void enableLinkButt(void);
 
 extern HWND g_hWnd;
